Beta version 1!!!
We have a full intro now! and it's GREAT!

There's a bunch of cool stuff going on here, I can't wait to tweak the project some more and add more cool levels as we go. I'm still a little intimidated by the ending but we'll see what we can get.